# apps/cluster-api-manager/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Package classification for clusterapimanager.",
    "title": "clusterapimanager ...\nbackend for pair",
    "contact": {
      "name": "Sharingio Pair",
      "email": "sharingio-pair@ii.coop"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.0.1"
  },
  "host": "localhost",
  "basePath": "/api",
  "paths": {
...
    "/instance": {
      "get": {
        "description": "List all instances",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "instance"
        ],
        "operationId": "listInstances",
        "responses": {
          "200": {
            "$ref": "#/responses/instanceList"
          },
          "500": {
            "$ref": "#/responses/failure"
          }
        }
      },
      "post": {
        "description": "creates an instance",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "instance"
        ],
        "operationId": "postInstance",
        "responses": {
          "200": {
            "$ref": "#/responses/instance"
          },
          "500": {
            "$ref": "#/responses/failure"
          }
        }
      },
      "delete": {
        "description": "delete an instance",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "instance"
        ],
        "operationId": "deleteInstance",
        "responses": {
          "200": {
            "$ref": "#/responses/instance"
          },
          "500": {
            "$ref": "#/responses/failure"
          }
        }
      }
    },
    "/instance/kubernetes": {
      "get": {
        "description": "List all Kubernetes instances",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "instance"
        ],
        "operationId": "listInstancesKubernetes",
        "responses": {
          "200": {
            "$ref": "#/responses/instanceList"
          },
          "500": {
            "$ref": "#/responses/failure"
          }
        }
      }
    },
    "/instance/kubernetes/{name}": {
      "get": {
        "description": "get a Kubernetes instance",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "instance"
        ],
        "operationId": "getInstanceKubernetes",
        "responses": {
          "200": {
            "$ref": "#/responses/instance"
          },
          "500": {
            "$ref": "#/responses/failure"
          }
        }
      },
      "delete": {
        "description": "delete a Kubernetes instance",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "instance"
        ],
        "operationId": "deleteInstanceKubernetes",
        "responses": {
          "200": {
            "$ref": "#/responses/instance"
          },
          "500": {
            "$ref": "#/responses/failure"
          }
        }
      }
    },
...
  },
  "definitions": {
    "ClusterStatus": {
      "description": "ClusterStatus defines the observed state of Cluster",
      "type": "object",
      "properties": {
        "conditions": {
          "$ref": "#/definitions/Conditions"
        },
        "controlPlaneInitialized": {
          "description": "ControlPlaneInitialized defines if the control plane has been initialized.\n+optional",
          "type": "boolean",
          "x-go-name": "ControlPlaneInitialized"
        },
        "controlPlaneReady": {
          "description": "ControlPlaneReady defines if the control plane is ready.\n+optional",
          "type": "boolean",
          "x-go-name": "ControlPlaneReady"
        },
        "failureDomains": {
          "$ref": "#/definitions/FailureDomains"
        },
        "failureMessage": {
          "description": "FailureMessage indicates that there is a fatal problem reconciling the\nstate, and will be set to a descriptive error message.\n+optional",
          "type": "string",
          "x-go-name": "FailureMessage"
        },
        "failureReason": {
          "$ref": "#/definitions/ClusterStatusError"
        },
        "infrastructureReady": {
          "description": "InfrastructureReady is the state of the infrastructure provider.\n+optional",
          "type": "boolean",
          "x-go-name": "InfrastructureReady"
        },
        "observedGeneration": {
          "description": "ObservedGeneration is the latest generation observed by the controller.\n+optional",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ObservedGeneration"
        },
        "phase": {
          "description": "Phase represents the current phase of cluster actuation.\nE.g. Pending, Running, Terminating, Failed etc.\n+optional",
          "type": "string",
          "x-go-name": "Phase"
        }
      },
      "x-go-package": "sigs.k8s.io/cluster-api/api/v1alpha3"
    },
    "ClusterStatusError": {
      "type": "string",
      "x-go-package": "sigs.k8s.io/cluster-api/errors"
    },
...
    "ConditionStatus": {
      "type": "string",
      "x-go-package": "k8s.io/api/core/v1"
    },
...
    "Conditions": {
      "type": "array",
      "title": "Conditions provide observations of the operational state of a Cluster API resource.",
      "items": {
        "$ref": "#/definitions/Condition"
      },
      "x-go-package": "sigs.k8s.io/cluster-api/api/v1alpha3"
    },
...
    "FailureDomains": {
      "type": "object",
      "title": "FailureDomains is a slice of FailureDomains.",
      "additionalProperties": {
        "$ref": "#/definitions/FailureDomainSpec"
      },
      "x-go-package": "sigs.k8s.io/cluster-api/api/v1alpha3"
    },
    "InstanceResourceStatus": {
      "description": "various status fields for an instance",
      "type": "object",
      "title": "InstanceResourceStatus ...",
      "properties": {
        "Cluster": {
          "$ref": "#/definitions/ClusterStatus"
        },
        "HumacsPod": {
          "$ref": "#/definitions/PodStatus"
        },
        "KubeadmControlPlane": {
          "$ref": "#/definitions/KubeadmControlPlaneStatus"
        },
        "MachineStatus": {
          "$ref": "#/definitions/MachineStatus"
        },
        "PacketMachineUID": {
          "type": "string"
        }
      },
      "x-go-package": "github.com/sharingio/pair/instances"
    },
    "InstanceSpec": {
      "description": "specification for an instance",
      "type": "object",
      "title": "InstanceSpec ...",
      "properties": {
        "facility": {
          "type": "string",
          "x-go-name": "Facility"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "nodeSize": {
          "type": "string",
          "x-go-name": "NodeSize"
        },
        "setup": {
          "$ref": "#/definitions/SetupSpec"
        },
        "type": {
          "$ref": "#/definitions/InstanceType"
        }
      },
      "x-go-package": "github.com/sharingio/pair/instances"
    },
    "InstanceStatus": {
      "description": "status fields",
      "type": "object",
      "title": "InstanceStatus ...",
      "properties": {
        "phase": {
          "$ref": "#/definitions/InstanceStatusPhase"
        },
        "resources": {
          "$ref": "#/definitions/InstanceResourceStatus"
        }
      },
      "x-go-package": "github.com/sharingio/pair/instances"
    },
    "InstanceStatusPhase": {
      "description": "Instance phase status definitions",
      "type": "string",
      "title": "InstanceStatusPhase ...",
      "x-go-package": "github.com/sharingio/pair/instances"
    },
    "InstanceType": {
      "description": "types of valid instances",
      "type": "string",
      "title": "InstanceType ...",
      "x-go-package": "github.com/sharingio/pair/instances"
    },
    "JSONResponseMetadata": {
      "description": "metadata fields in responses",
      "type": "object",
      "title": "JSONResponseMetadata ...",
      "properties": {
        "requestId": {
          "type": "string",
          "x-go-name": "RequestId"
        },
        "response": {
          "type": "string",
          "x-go-name": "Response"
        },
        "selfLink": {
          "type": "string",
          "x-go-name": "URL"
        },
        "timestamp": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Timestamp"
        },
        "version": {
          "type": "string",
          "x-go-name": "Version"
        }
      },
      "x-go-package": "github.com/sharingio/pair/types"
    },
...
    "PodCondition": {
      "type": "object",
      "title": "PodCondition contains details for the current condition of this pod.",
      "properties": {
        "lastProbeTime": {
          "$ref": "#/definitions/Time"
        },
        "lastTransitionTime": {
          "$ref": "#/definitions/Time"
        },
        "message": {
          "description": "Human-readable message indicating details about last transition.\n+optional",
          "type": "string",
          "x-go-name": "Message"
        },
        "reason": {
          "description": "Unique, one-word, CamelCase reason for the condition's last transition.\n+optional",
          "type": "string",
          "x-go-name": "Reason"
        },
        "status": {
          "$ref": "#/definitions/ConditionStatus"
        },
        "type": {
          "$ref": "#/definitions/PodConditionType"
        }
      },
      "x-go-package": "k8s.io/api/core/v1"
    },
...
    "PodPhase": {
      "type": "string",
      "title": "PodPhase is a label for the condition of a pod at the current time.",
      "x-go-package": "k8s.io/api/core/v1"
    },
    "PodQOSClass": {
      "type": "string",
      "title": "PodQOSClass defines the supported qos classes of Pods.",
      "x-go-package": "k8s.io/api/core/v1"
    },
    "PodStatus": {
      "description": "PodStatus represents information about the status of a pod. Status may trail the actual\nstate of a system, especially if the node that hosts the pod cannot contact the control\nplane.",
      "type": "object",
      "properties": {
        "conditions": {
          "description": "Current service state of pod.\nMore info: https://kubernetes.io/docs/concepts/workloads/pods/pod-lifecycle#pod-conditions\n+optional\n+patchMergeKey=type\n+patchStrategy=merge",
          "type": "array",
          "items": {
            "$ref": "#/definitions/PodCondition"
          },
          "x-go-name": "Conditions"
        },
        "containerStatuses": {
          "description": "The list has one entry per container in the manifest. Each entry is currently the output\nof `docker inspect`.\nMore info: https://kubernetes.io/docs/concepts/workloads/pods/pod-lifecycle#pod-and-container-status\n+optional",
          "type": "array",
          "items": {
            "$ref": "#/definitions/ContainerStatus"
          },
          "x-go-name": "ContainerStatuses"
        },
        "ephemeralContainerStatuses": {
          "description": "Status for any ephemeral containers that have run in this pod.\nThis field is alpha-level and is only populated by servers that enable the EphemeralContainers feature.\n+optional",
          "type": "array",
          "items": {
            "$ref": "#/definitions/ContainerStatus"
          },
          "x-go-name": "EphemeralContainerStatuses"
        },
        "hostIP": {
          "description": "IP address of the host to which the pod is assigned. Empty if not yet scheduled.\n+optional",
          "type": "string",
          "x-go-name": "HostIP"
        },
        "initContainerStatuses": {
          "description": "The list has one entry per init container in the manifest. The most recent successful\ninit container will have ready = true, the most recently started container will have\nstartTime set.\nMore info: https://kubernetes.io/docs/concepts/workloads/pods/pod-lifecycle#pod-and-container-status",
          "type": "array",
          "items": {
            "$ref": "#/definitions/ContainerStatus"
          },
          "x-go-name": "InitContainerStatuses"
        },
        "message": {
          "description": "A human readable message indicating details about why the pod is in this condition.\n+optional",
          "type": "string",
          "x-go-name": "Message"
        },
        "nominatedNodeName": {
          "description": "nominatedNodeName is set only when this pod preempts other pods on the node, but it cannot be\nscheduled right away as preemption victims receive their graceful termination periods.\nThis field does not guarantee that the pod will be scheduled on this node. Scheduler may decide\nto place the pod elsewhere if other nodes become available sooner. Scheduler may also decide to\ngive the resources on this node to a higher priority pod that is created after preemption.\nAs a result, this field may be different than PodSpec.nodeName when the pod is\nscheduled.\n+optional",
          "type": "string",
          "x-go-name": "NominatedNodeName"
        },
        "phase": {
          "$ref": "#/definitions/PodPhase"
        },
        "podIP": {
          "description": "IP address allocated to the pod. Routable at least within the cluster.\nEmpty if not yet allocated.\n+optional",
          "type": "string",
          "x-go-name": "PodIP"
        },
        "podIPs": {
          "description": "podIPs holds the IP addresses allocated to the pod. If this field is specified, the 0th entry must\nmatch the podIP field. Pods may be allocated at most 1 value for each of IPv4 and IPv6. This list\nis empty if no IPs have been allocated yet.\n+optional\n+patchStrategy=merge\n+patchMergeKey=ip",
          "type": "array",
          "items": {
            "$ref": "#/definitions/PodIP"
          },
          "x-go-name": "PodIPs"
        },
        "qosClass": {
          "$ref": "#/definitions/PodQOSClass"
        },
        "reason": {
          "description": "A brief CamelCase message indicating details about why the pod is in this state.\ne.g. 'Evicted'\n+optional",
          "type": "string",
          "x-go-name": "Reason"
        },
        "startTime": {
          "$ref": "#/definitions/Time"
        }
      },
      "x-go-package": "k8s.io/api/core/v1"
    },
    "SetupSpec": {
      "description": "fields for provisioning an instance",
      "type": "object",
      "title": "SetupSpec ...",
      "properties": {
        "baseDNSName": {
          "type": "string",
          "x-go-name": "BaseDNSName"
        },
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "env": {
          "type": "array",
          "items": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "x-go-name": "Env"
        },
        "fullname": {
          "type": "string",
          "x-go-name": "Fullname"
        },
        "githubOAuthToken": {
          "type": "string",
          "x-go-name": "GitHubOAuthToken"
        },
        "guests": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Guests"
        },
        "repos": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Repos"
        },
        "timezone": {
          "type": "string",
          "x-go-name": "Timezone"
        },
        "user": {
          "type": "string",
          "x-go-name": "User"
        }
      },
      "x-go-package": "github.com/sharingio/pair/types"
    },
...
  },
  "responses": {
    "failure": {
      "description": "JSONFailure ...\ngeneric JSON for failure",
      "schema": {
        "$ref": "#/definitions/JSONResponseMetadata"
      },
      "headers": {
        "metadata": {}
      }
    },
    "instance": {
      "description": "Instance ...\ngeneric instance",
      "schema": {
        "$ref": "#/definitions/InstanceStatus"
      },
      "headers": {
        "metadata": {},
        "spec": {},
        "status": {}
      }
    },
...
    "instanceList": {
      "description": "InstanceList ...\ngeneric instance list",
      "schema": {
        "$ref": "#/definitions/JSONResponseMetadata"
      },
      "headers": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/InstanceSpec"
          }
        },
        "metadata": {}
      }
    }
  }
}